Move tests into corresponding folders
